修改-角色及资源业务逻辑

+++ user/src/main/java/com/twj/spirngbasics/user/service/UserRoleService.java
@@ -2,10 +2,13 @@ package com.twj.spirngbasics.user.service;
 
 import cn.hutool.core.util.IdUtil;
 import com.github.pagehelper.PageHelper;
+import com.github.pagehelper.PageInfo;
 import com.twj.spirngbasics.server.dto.PageDto;
+import com.twj.spirngbasics.server.manage.RedisManage;
 import com.twj.spirngbasics.server.util.CopyUtils;
 import com.twj.spirngbasics.server.util.ValidatorUtils;
 import com.twj.spirngbasics.user.dto.UserRoleDto;
+import com.twj.spirngbasics.user.entity.UserResource;
 import com.twj.spirngbasics.user.entity.UserRole;
 import com.twj.spirngbasics.user.entity.UserRoleResource;
 import com.twj.spirngbasics.user.entity.UserRoleUser;
@@ -13,14 +16,14 @@ import com.twj.spirngbasics.user.mapper.*;
 import lombok.extern.slf4j.Slf4j;
 import org.mybatis.dynamic.sql.SqlBuilder;
 import org.mybatis.dynamic.sql.delete.render.DeleteStatementProvider;
-import org.mybatis.dynamic.sql.insert.render.MultiRowInsertStatementProvider;
 import org.mybatis.dynamic.sql.render.RenderingStrategies;
 import org.mybatis.dynamic.sql.select.render.SelectStatementProvider;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.StringUtils;
 
 import javax.annotation.Resource;
-import java.util.ArrayList;
-import java.util.List;
+import java.util.*;
 
 /**
  * @作者: Jun
@@ -41,39 +44,49 @@ public class UserRoleService {
     @Resource
     private UserRoleUserMapper userRoleUserMapper;
 
+    @Resource
+    private UserResourceMapper userResourceMapper;
+
     public UserRoleDto findById(String id) {
         ValidatorUtils.require(id, "参数不能为空");
         return CopyUtils.copy(userRoleMapper.selectByPrimaryKey(id).get(), UserRoleDto.class);
     }
 
-    /**
-     * 列表查询
-     */
     public void list(PageDto pageDto) {
         PageHelper.startPage(pageDto.getPage(), pageDto.getSize());
-        SelectStatementProvider selectStatement = SqlBuilder.select(userRoleMapper.selectList)
+        SelectStatementProvider selectStatement = SqlBuilder.select(UserRoleMapper.selectList)
                 .from(UserRoleDynamicSqlSupport.userRole)
                 .build()
                 .render(RenderingStrategies.MYBATIS3);
-        List<UserRoleDto> userRoleDtoList = CopyUtils.copyList(userRoleMapper.selectMany(selectStatement), UserRoleDto.class);
-        pageDto.setTotal(userRoleDtoList.size());
-        pageDto.setList(userRoleDtoList);
+        List<UserRole> list = userRoleMapper.selectMany(selectStatement);
+        PageInfo pageInfo = new PageInfo(list);
+        pageDto.setTotal(pageInfo.getTotal());
+        pageDto.setList(CopyUtils.copyList(list, UserRoleDto.class));
     }
 
     /**
      * 新增
      */
+    @Transactional
     public void insert(UserRole userRole) {
         userRole.insert();
         userRoleMapper.insert(userRole);
+        //保存角色资源
+        saveResource(userRole);
+        userRole.setCompanyId(null);
     }
 
     /**
      * 更新
      */
+    @Transactional
     public void update(UserRole userRole) {
         userRole.update();
+        //为了保证安全，后台创建的不允许修改companyid
+        userRole.setCompanyId(null);
         userRoleMapper.updateByPrimaryKeySelective(userRole);
+        saveResource(userRole);
+        RedisManage.delRoleResource(userRole.getId());
     }
 
     /**
@@ -87,33 +100,31 @@ public class UserRoleService {
     /**
      * 按角色保存资源
      */
-    public void saveResource(UserRoleDto roleDto) {
-        String roleId = roleDto.getId();
-        List<String> resourceIds = roleDto.getResourceIds();
+    public void saveResource(UserRole role) {
+        String roleId = role.getId();
+        if (role.getResourceIds() == null || role.getResourceIds().size() == 0) {
+            return;
+        }
+        List<String> resourceIds = role.getResourceIds();
         // 清空库中所有的当前角色下的记录
         DeleteStatementProvider deleteStatement = SqlBuilder.deleteFrom(UserRoleResourceDynamicSqlSupport.userRoleResource)
-                .where(UserRoleResourceDynamicSqlSupport.roleId, SqlBuilder.isNotEqualToWhenPresent(roleId))
+                .where(UserRoleResourceDynamicSqlSupport.roleId, SqlBuilder.isEqualTo(roleId))
                 .build()
                 .render(RenderingStrategies.MYBATIS3);
-        userRoleMapper.delete(deleteStatement);
+        userRoleResourceMapper.delete(deleteStatement);
 
 
         // 保存角色资源
         List<UserRoleResource> insertList = new ArrayList<>();
         for (int i = 0; i < resourceIds.size(); i++) {
             UserRoleResource roleResource = new UserRoleResource();
-            roleResource.setId(IdUtil.simpleUUID());
             roleResource.setRoleId(roleId);
             roleResource.setResourceId(resourceIds.get(i));
+            roleResource.insert();
             insertList.add(roleResource);
         }
 
-        MultiRowInsertStatementProvider<UserRoleResource> multiRowInsert = SqlBuilder.insertMultiple(insertList)
-                .into(UserRoleResourceDynamicSqlSupport.userRoleResource)
-                .build()
-                .render(RenderingStrategies.MYBATIS3);
-
-        userRoleResourceMapper.insertMultiple(multiRowInsert);
+        userRoleResourceMapper.insertMultiple(insertList);
     }
 
     /**
@@ -181,4 +192,24 @@ public class UserRoleService {
         return userIdList;
     }
 
+
+    public Map<String, Object> getUserResourceList(String roleId) {
+        SelectStatementProvider select = SqlBuilder.select(userResourceMapper.selectList)
+                .from(UserResourceDynamicSqlSupport.userResource)
+                .where(UserResourceDynamicSqlSupport.id, SqlBuilder.isIn(
+                        SqlBuilder.select(UserRoleResourceDynamicSqlSupport.resourceId)
+                                .from(UserRoleResourceDynamicSqlSupport.userRoleResource)
+                                .where(UserRoleResourceDynamicSqlSupport.roleId, SqlBuilder.isEqualTo(roleId)
+                                )))
+                .build()
+                .render(RenderingStrategies.MYBATIS3);
+        List<UserResource> resources = userResourceMapper.selectMany(select);
+        Map<String, Object> resourceMap = new HashMap<>();
+        for (UserResource resource : resources) {
+            if (!StringUtils.isEmpty(resource.getRequest())) {
+                resourceMap.put(resource.getRequest(), userResourceMapper.selectByPrimaryKey(resource.getParent()).get().getName() + "-" + resource.getName());
+            }
+        }
+        return resourceMap;
+    }
 }
